backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
timesfm[torch]
google-auth-oauthlib
google-cloud-bigquery
google-cloud-storage
db-dtypes
pandas-gbq
